Use KUBECONFIG to identify k8s configuration path, fallback otherwise
kubectl uses this environment variable to find its configuration file,
but uses the one in the default ($HOME) location otherwise.

Click now does the same, which makes it comply with what people would
expect.

The reason I am making this PR is that we do have a tool which
configures K8s on the fly, and produces a k8s configuration in a
temporary location.